ioctl's int->long

+++ main.c
@@ -398,7 +398,8 @@ ssize_t scull_write(struct file *filp, const char __user *buf, size_t count,
  * The ioctl() implementation
  */
 
-int scull_ioctl(/*struct inode *inode,*/ struct file *filp, unsigned int cmd, unsigned long arg)
+long scull_ioctl(struct file *filp, unsigned int cmd, unsigned long arg)
+/*int scull_ioctl(struct inode *inode, struct file *filp, unsigned int cmd, unsigned long arg)*/
 {
 
 	int err = 0, tmp;
@@ -556,13 +557,13 @@ loff_t scull_llseek(struct file *filp, loff_t off, int whence)
 
 
 struct file_operations scull_fops = {
-	.owner =    THIS_MODULE,
-	.llseek =   scull_llseek,
-	.read =     scull_read,
-	.write =    scull_write,
-	.unlocked_ioctl =    scull_ioctl,
-	.open =     scull_open,
-	.release =  scull_release,
+	.owner =    		THIS_MODULE,
+	.llseek =   		scull_llseek,
+	.read =     		scull_read,
+	.write =    		scull_write,
+	.unlocked_ioctl =    	scull_ioctl,
+	.open =     		scull_open,
+	.release =  		scull_release,
 };
 
 /*
